requirements.txt
elevenlabs>=1.0.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
groq==0.5.0
python-dotenv==1.0.0
sentence-transformers
torch
scikit-learn
word2number==1.1
numpy
pydantic